Fix #11: "A"SR Freeze when scrolling past USER1

+++ software/o_c_REV/HEM_ASR.ino
@@ -106,7 +106,7 @@ private:
     int cursor;
     RingBufferManager *buffer_m = buffer_m->get();
     braids::Quantizer quantizer;
-    byte scale;
+    int scale;
     
     void DrawInterface() {
         // Show Link icon if linked with another ASR
